refactor(frontend): extract steerMessage hook to keep useAgent under 1000 lines
