+++ frontend/src/hooks/useAgent.ts
@@ -3,7 +3,7 @@
  * Provides agent communication, message management, and SSE streaming
  */
 
-import { useState, useCallback, useRef, useEffect } from "react";
+import { useState, useCallback, useRef, useEffect, useMemo } from "react";
 import toast from "react-hot-toast";
 import i18n from "../i18n";
 import type {
@@ -14,7 +14,6 @@ import type {
   MessageAttachment,
 } from "../types";
 import { sessionApi, type BackendSession } from "../services/api";
-import { buildSteerUserMessage } from "../utils/steerMessages";
 import { authenticatedRequest } from "../services/api/authenticatedRequest";
 import { API_BASE } from "../services/api/config";
 import { feedbackApi } from "../services/api/feedback";
@@ -45,6 +44,7 @@ import {
   type SSEConnectionContext,
 } from "./useAgent/sseConnection";
 import { createOptimisticMessagesForSend } from "./useAgent/optimisticMessages";
+import { createSteerMessage } from "./useAgent/steerMessage";
 import { getValidAccessToken } from "../services/api/tokenManager";
 import { resolveRunEnabledSkills } from "./useAgent/runSkillOverrides";
 import { planGoalSubmission } from "./useAgent/goalCommands";
@@ -858,25 +858,10 @@ export function useAgent(options?: UseAgentOptions): UseAgentReturn {
     }
   }, [options]);
 
-  // Codex 式运行中插话：消息进入后端队列，当前步骤后注入；先乐观展示
-  const steerMessage = useCallback(async (content: string) => {
-    const text = content.trim();
-    const currentSessionId = sessionIdRef.current;
-    if (!text || !currentSessionId) return;
-
-    const optimistic = buildSteerUserMessage({
-      previousCount: 0,
-      content: text,
-    });
-    setMessages((prev) => [...prev, optimistic]);
-    try {
-      await sessionApi.steer(currentSessionId, text);
-    } catch (error) {
-      console.error("[steerMessage] Failed to steer session:", error);
-      setMessages((prev) => prev.filter((m) => m.id !== optimistic.id));
-      setError(i18n.t("chat.steerFailed", "插话发送失败，请稍后重试"));
-    }
-  }, []);
+  const steerMessage = useMemo(
+    () => createSteerMessage({ sessionIdRef, setMessages, setError }),
+    [],
+  );
 
   const clearMessages = useCallback(() => {
     loadHistoryRequestIdRef.current += 1;
